:poop: 删除台帐逻辑+判断告警策略

+++ ems-module-power/ems-module-power-biz/src/main/java/cn/bitlinks/ems/module/power/service/standingbook/StandingbookServiceImpl.java
@@ -2,6 +2,7 @@ package cn.bitlinks.ems.module.power.service.standingbook;
 
 import cn.bitlinks.ems.framework.common.util.object.BeanUtils;
 import cn.bitlinks.ems.framework.mybatis.core.query.LambdaQueryWrapperX;
+import cn.bitlinks.ems.module.acquisition.api.quartz.QuartzApi;
 import cn.bitlinks.ems.module.power.controller.admin.deviceassociationconfiguration.vo.AssociationData;
 import cn.bitlinks.ems.module.power.controller.admin.deviceassociationconfiguration.vo.StandingbookWithAssociations;
 import cn.bitlinks.ems.module.power.controller.admin.standingbook.attribute.vo.StandingbookAttributeSaveReqVO;
@@ -12,16 +13,12 @@ import cn.bitlinks.ems.module.power.dal.dataobject.measurementdevice.Measurement
 import cn.bitlinks.ems.module.power.dal.dataobject.standingbook.StandingbookDO;
 import cn.bitlinks.ems.module.power.dal.dataobject.standingbook.StandingbookLabelInfoDO;
 import cn.bitlinks.ems.module.power.dal.dataobject.standingbook.acquisition.StandingbookAcquisitionDO;
-import cn.bitlinks.ems.module.power.dal.dataobject.standingbook.acquisition.StandingbookAcquisitionDetailDO;
 import cn.bitlinks.ems.module.power.dal.dataobject.standingbook.attribute.StandingbookAttributeDO;
-import cn.bitlinks.ems.module.power.dal.dataobject.standingbook.tmpl.StandingbookTmplDaqAttrDO;
 import cn.bitlinks.ems.module.power.dal.dataobject.standingbook.type.StandingbookTypeDO;
 import cn.bitlinks.ems.module.power.dal.mysql.measurementassociation.MeasurementAssociationMapper;
 import cn.bitlinks.ems.module.power.dal.mysql.measurementdevice.MeasurementDeviceMapper;
 import cn.bitlinks.ems.module.power.dal.mysql.standingbook.StandingbookLabelInfoMapper;
 import cn.bitlinks.ems.module.power.dal.mysql.standingbook.StandingbookMapper;
-import cn.bitlinks.ems.module.power.dal.mysql.standingbook.acquisition.StandingbookAcquisitionDetailMapper;
-import cn.bitlinks.ems.module.power.dal.mysql.standingbook.acquisition.StandingbookAcquisitionMapper;
 import cn.bitlinks.ems.module.power.dal.mysql.standingbook.attribute.StandingbookAttributeMapper;
 import cn.bitlinks.ems.module.power.dal.mysql.standingbook.templ.StandingbookTmplDaqAttrMapper;
 import cn.bitlinks.ems.module.power.dal.mysql.standingbook.type.StandingbookTypeMapper;
@@ -29,18 +26,14 @@ import cn.bitlinks.ems.module.power.enums.CommonConstants;
 import cn.bitlinks.ems.module.power.enums.ErrorCodeConstants;
 import cn.bitlinks.ems.module.power.enums.standingbook.StandingbookTypeTopEnum;
 import cn.bitlinks.ems.module.power.service.standingbook.acquisition.StandingbookAcquisitionService;
-import cn.bitlinks.ems.module.power.service.labelconfig.LabelConfigService;
-import cn.bitlinks.ems.module.power.service.standingbook.acquisition.StandingbookAcquisitionService;
 import cn.bitlinks.ems.module.power.service.standingbook.attribute.StandingbookAttributeService;
+import cn.bitlinks.ems.module.power.service.warningstrategy.WarningStrategyService;
 import cn.hutool.core.collection.CollUtil;
 import cn.hutool.json.JSONObject;
 import cn.hutool.json.JSONUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.toolkit.StringPool;
 import com.baomidou.mybatisplus.core.toolkit.StringUtils;
-import org.springframework.context.annotation.Lazy;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -81,7 +74,11 @@ public class StandingbookServiceImpl implements StandingbookService {
     private MeasurementDeviceMapper measurementDeviceMapper;
     @Resource
     private MeasurementAssociationMapper measurementAssociationMapper;
-
+    @Resource
+    @Lazy
+    private WarningStrategyService warningStrategyService;
+    @Resource
+    private QuartzApi quartzApi;
     @Lazy
     @Resource
     private StandingbookAcquisitionService standingbookAcquisitionService;
@@ -276,7 +273,7 @@ public class StandingbookServiceImpl implements StandingbookService {
         standingbook.setTypeId(Long.valueOf(updateReqVO.get("typeId")));
         standingbook.setId(Long.valueOf(updateReqVO.get("id")));
         // 修改标签信息 先删后增
-        if(StringUtils.isNotEmpty(updateReqVO.get(ATTR_LABEL_INFO))){
+        if (StringUtils.isNotEmpty(updateReqVO.get(ATTR_LABEL_INFO))) {
             standingbookLabelInfoMapper.delete(new LambdaQueryWrapper<StandingbookLabelInfoDO>().eq(StandingbookLabelInfoDO::getStandingbookId, standingbook.getId()));
             createLabelInfoList(updateReqVO.get(ATTR_LABEL_INFO), standingbook.getId());
         }
@@ -303,45 +300,50 @@ public class StandingbookServiceImpl implements StandingbookService {
         }
 
     }
+
     @Transactional
     @Override
     public void deleteStandingbookBatch(List<Long> ids) {
-        if(CollUtil.isEmpty(ids)){
+        if (CollUtil.isEmpty(ids)) {
             return;
         }
         // 如果存在关联关系，则不删除台账
         Long count = measurementDeviceMapper.selectCount(new LambdaQueryWrapper<MeasurementDeviceDO>()
-                .in(MeasurementDeviceDO::getDeviceId,ids)
-                .or().in(MeasurementDeviceDO::getMeasurementInstrumentId,ids));
-        if(count >0 ){
+                .in(MeasurementDeviceDO::getDeviceId, ids)
+                .or().in(MeasurementDeviceDO::getMeasurementInstrumentId, ids));
+        if (count > 0) {
             throw exception(ErrorCodeConstants.STANDINGBOOK_ASSOCIATION_EXISTS);
         }
         count = measurementAssociationMapper.selectCount(new LambdaQueryWrapper<MeasurementAssociationDO>()
-                .in(MeasurementAssociationDO::getMeasurementId,ids)
-                .or().in(MeasurementAssociationDO::getMeasurementInstrumentId,ids));
-        if(count>0){
+                .in(MeasurementAssociationDO::getMeasurementId, ids)
+                .or().in(MeasurementAssociationDO::getMeasurementInstrumentId, ids));
+        if (count > 0) {
             throw exception(ErrorCodeConstants.STANDINGBOOK_ASSOCIATION_EXISTS);
         }
         // 查询存在启用的数采关联
         List<StandingbookAcquisitionDO> standingbookAcquisitionList =
                 standingbookAcquisitionService.queryListByStandingbookIds(ids);
-        if(CollUtil.isNotEmpty(standingbookAcquisitionList)){
+        if (CollUtil.isNotEmpty(standingbookAcquisitionList)) {
             throw exception(ErrorCodeConstants.STANDINGBOOK_ACQUISITION_EXISTS);
+        }
+        // 查询存在告警配置
+        if (warningStrategyService.existsByStandingbookIds(ids)) {
+            throw exception(ErrorCodeConstants.STANDINGBOOK_REL_STRATEGY);
         }
 
         // 删除
         standingbookMapper.deleteByIds(ids);
         // 删除标签信息
         standingbookLabelInfoMapper.delete(new LambdaQueryWrapperX<StandingbookLabelInfoDO>()
-                .inIfPresent(StandingbookLabelInfoDO::getStandingbookId,ids));
+                .inIfPresent(StandingbookLabelInfoDO::getStandingbookId, ids));
         // 删除属性
         standingbookAttributeMapper.delete(new LambdaQueryWrapperX<StandingbookAttributeDO>()
-                .inIfPresent(StandingbookAttributeDO::getStandingbookId,ids));
+                .inIfPresent(StandingbookAttributeDO::getStandingbookId, ids));
 
         // 删除数采关联
         standingbookAcquisitionService.deleteByStandingbookIds(ids);
-        // 删除数采的任务 todo
-
+        // 删除数采的任务
+        quartzApi.deleteJob(ids);
 
     }
 
@@ -381,7 +383,7 @@ public class StandingbookServiceImpl implements StandingbookService {
         List<Long> energyTypeIds = new ArrayList<>();
         if (StringUtils.isNotEmpty(energy)) {
             energyTypeIds = standingbookTmplDaqAttrMapper.selectSbTypeIdsByEnergyId(Long.valueOf(energy));
-            if(CollUtil.isEmpty(energyTypeIds)){
+            if (CollUtil.isEmpty(energyTypeIds)) {
                 return Collections.emptyList();
             }
         }
